fix(fixed folder archiiteture and renamed files): fixed folder architecture and renamed files
