+++ src/App.tsx
@@ -1,8 +1,8 @@
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 
-import scss from './App.module.scss';
+import Home from './pages/home/home';
 
-import Home from './pages/home/';
+import scss from './App.module.scss';
 
 function App() {
   return (
